Move MultiCouplingThreeSolvers.
