+++ tests/serial/multi-coupling/helpers.cpp
@@ -7,6 +7,113 @@
 
 using namespace precice;
 
+// Simple case of A <==> B <==> C
+void multiCouplingThreeSolvers(const std::string configFile, const TestContext &context)
+{
+  Eigen::Vector2d coordOneA{0.0, 0.0};
+  std::string     writeIterCheckpoint(constants::actionWriteIterationCheckpoint());
+  std::string     readIterCheckpoint(constants::actionReadIterationCheckpoint());
+
+  double valueA = 1.0;
+  double valueB = 2.0;
+  double valueC = 3.0;
+
+  if (context.isNamed("SolverA")) {
+    SolverInterface cplInterface("SolverA", configFile, 0, 1);
+    const int       meshID   = cplInterface.getMeshID("MeshA");
+    int             vertexID = cplInterface.setMeshVertex(meshID, coordOneA.data());
+    int             dataABID = cplInterface.getDataID("DataAB", meshID);
+    int             dataBAID = cplInterface.getDataID("DataBA", meshID);
+
+    double maxDt = cplInterface.initialize();
+    double valueRead;
+
+    BOOST_TEST(cplInterface.isCouplingOngoing());
+    while (cplInterface.isCouplingOngoing()) {
+      cplInterface.writeScalarData(dataABID, vertexID, valueA);
+      if (cplInterface.isActionRequired(writeIterCheckpoint)) {
+        cplInterface.markActionFulfilled(writeIterCheckpoint);
+      }
+
+      cplInterface.advance(maxDt);
+
+      if (cplInterface.isActionRequired(readIterCheckpoint)) {
+        cplInterface.markActionFulfilled(readIterCheckpoint);
+      }
+      cplInterface.readScalarData(dataBAID, vertexID, valueRead);
+    }
+
+    BOOST_TEST(valueRead == valueB);
+
+    cplInterface.finalize();
+  } else if (context.isNamed("SolverB")) {
+    SolverInterface cplInterface("SolverB", configFile, 0, 1);
+    const int       meshID1   = cplInterface.getMeshID("MeshB1");
+    const int       meshID2   = cplInterface.getMeshID("MeshB2");
+    int             vertexID1 = cplInterface.setMeshVertex(meshID1, coordOneA.data());
+    int             vertexID2 = cplInterface.setMeshVertex(meshID2, coordOneA.data());
+    int             dataABID  = cplInterface.getDataID("DataAB", meshID1);
+    int             dataBAID  = cplInterface.getDataID("DataBA", meshID1);
+    int             dataCBID  = cplInterface.getDataID("DataCB", meshID2);
+    int             dataBCID  = cplInterface.getDataID("DataBC", meshID2);
+
+    double maxDt = cplInterface.initialize();
+    double valueReadA, valueReadC;
+
+    BOOST_TEST(cplInterface.isCouplingOngoing());
+    while (cplInterface.isCouplingOngoing()) {
+      cplInterface.writeScalarData(dataBAID, vertexID1, valueB);
+      cplInterface.writeScalarData(dataBCID, vertexID2, valueB);
+      if (cplInterface.isActionRequired(writeIterCheckpoint)) {
+        cplInterface.markActionFulfilled(writeIterCheckpoint);
+      }
+
+      cplInterface.advance(maxDt);
+
+      if (cplInterface.isActionRequired(readIterCheckpoint)) {
+        cplInterface.markActionFulfilled(readIterCheckpoint);
+      }
+      cplInterface.readScalarData(dataABID, vertexID1, valueReadA);
+      cplInterface.readScalarData(dataCBID, vertexID2, valueReadC);
+    }
+
+    BOOST_TEST(valueReadA == 1.0);
+    BOOST_TEST(valueReadC == 3.0);
+
+    cplInterface.finalize();
+
+  } else {
+    SolverInterface cplInterface("SolverC", configFile, 0, 1);
+    const int       meshID   = cplInterface.getMeshID("MeshC");
+    int             vertexID = cplInterface.setMeshVertex(meshID, coordOneA.data());
+    int             dataCBID = cplInterface.getDataID("DataCB", meshID);
+    int             dataBCID = cplInterface.getDataID("DataBC", meshID);
+
+    double maxDt = cplInterface.initialize();
+    double valueRead;
+
+    BOOST_TEST(cplInterface.isCouplingOngoing());
+    while (cplInterface.isCouplingOngoing()) {
+
+      cplInterface.writeScalarData(dataCBID, vertexID, valueC);
+      if (cplInterface.isActionRequired(writeIterCheckpoint)) {
+        cplInterface.markActionFulfilled(writeIterCheckpoint);
+      }
+
+      cplInterface.advance(maxDt);
+
+      if (cplInterface.isActionRequired(readIterCheckpoint)) {
+        cplInterface.markActionFulfilled(readIterCheckpoint);
+      }
+      cplInterface.readScalarData(dataBCID, vertexID, valueRead);
+    }
+
+    BOOST_TEST(valueRead == 2.0);
+
+    cplInterface.finalize();
+  }
+}
+
 void multiCouplingFourSolvers(const std::string configFile, const TestContext &context)
 {
   Eigen::Vector2d coordOneA{0.0, 0.0};
