+++ tests/serial/multi-coupling/helpers.hpp
@@ -7,6 +7,8 @@
 using namespace precice;
 using precice::testing::TestContext;
 
+void multiCouplingThreeSolvers(const std::string configFile, const TestContext &context);
+
 void multiCouplingFourSolvers(const std::string configFile, const TestContext &context);
 
 #endif
